Refactor Korro tasks: replace `korroApply` with `korroGenerate` and `korro` for improved caching and clarity
- Deprecated `korroApply` in favor of a streamlined task structure with `korroGenerate` and `korro`.
- Introduced `korroGenerate` (cacheable) for isolated markdown generation into `build/korro/docs/` without mutating sources.
- Updated `korro` to depend on `korroGenerate` and handle applying generated docs onto the source tree.
- Updated documentation, integration tests, and example commands to reflect task changes.
- Enhanced task descriptions, clarified CI usage, and ensured safe regeneration workflows.
